Screen Resolution adjustment Level 1

+++ Bookshelf_Short.pde
@@ -2,19 +2,19 @@ import java.awt.*;
 
 class Bookshelf_Short extends Object {
   
-  Bookshelf_Short(int x) {
+  Bookshelf_Short(float x) {
     
     super(x);
     image = loadImage(sketchPath() + "/images/objects/bookshelf_short.png");
     platform = false;
     hitbox = new Rectangle[1];
-    hitbox[0] = new Rectangle(x+11, y+696, 175, 206);
+    hitbox[0] = new Rectangle((int)x+11, (int)y+696, 175, 206);
         
   }
   
   void draw() {
     image(image, x, y+516, 192, 384);
-    hitbox[0].setLocation(x+11, y+696);
+    hitbox[0].setLocation((int)x+11, (int)y+696);
   }
   
   void toggle() {
