+++ Chair.pde
@@ -2,19 +2,19 @@ import java.awt.*;
 
 class Chair extends Object {
   
-  Chair(int x) {
+  Chair(float x) {
     
     super(x);
     image = loadImage(sketchPath() + "/images/objects/chair.png");
     platform = false;
     hitbox = new Rectangle[1];
-    hitbox[0] = new Rectangle(x+25, y+846, 83, 54);
+    hitbox[0] = new Rectangle((int)x+25, (int)y+846, 83, 54);
         
   }
   
   void draw() {
     image(image, x, y+772, 128, 128);
-    hitbox[0].setLocation(x+25, y+846);
+    hitbox[0].setLocation((int)x+25, (int)y+846);
   }
   
   void toggle() {
